+++ src/map/map-max-size/map-max-size.decorator.ts
@@ -8,7 +8,7 @@ export const MAP_MAX_SIZE = 'mapMaxSize'
 
 /**
  * Checks if the given value is a [Map](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map)
- * with a size less than or equal to `maximum`..
+ * with a size less than or equal to `maximum`.
  *
  * #### Example
  * ```typescript