+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   `ArraySize`: Added new validator.
+-   `MapSize`: Added new validator.
+-   `SetSize`: Added new validator.
+
 ## [1.0.0] - 2022-04-03
 
 ### Chore